Huge Update as well as Google Play Store!
~=~ More Monsters show up ~=~
Stickfigure (Rage)
Ghost (Block)
~=~ New Store! ~=~
you can save inside as well as buy upgrade(s)
~=~ New Player Skills! ~=~
scrach (unlocked @ lv1)
~=~And more!~=~

~=~ Moving to google play! ~=~
with us moving to the google play store we add on more content adding
new monsters and new player skills and so much more!

~=~ New Script ~=~
we use to have only two scripts runing the game but _core needs some "me
time" and so we got one more to join the fun _PlayerAbl who handles all
the players skills and more it will be the script for all things skills
and talents

~=~ Beta Announcement! ~=~
We are moving from devleopment to Alpha!
With all these new things it's time we show the world how awesome we are
to join this Alpha just go to the google play store and apply!